Move locators in utils
